v2.0.0b4: Fold tiferet.testing into core package, revert separate tiferet-testing build
- Add tiferet/testing/ subpackage (domain.py, mappers.py, hooks.py, __init__.py)
- Rewrite all tiferet_testing imports to tiferet.testing across tests/
- Update pyproject.toml: add tiferet.testing to packages, drop tiferet-testing dep
- Revert workflow to single tiferet package publish
- Remove tiferet_testing/ directory and tiferet/tests_int/
