+++ .github/workflows/python-publish.yml
@@ -1,20 +1,18 @@
-# Unified workflow: builds and publishes Tiferet and Tiferet-Testing to PyPI.
+# Unified workflow: builds and publishes Tiferet to PyPI.
 #
 # Trigger:
 #   - Tag push matching 'v*' (e.g. v2.0.0b4, v2.0.0).
 #
 # Flow:
 #   1. test   — runs pytest on 3.10, 3.11, 3.12.
-#   2. build  — validates version/tag match, builds both packages into
-#               separate artifacts, runs twine check.
-#   3. publish — per-package publish jobs, each with a dedicated GitHub
-#                environment mapped 1:1 to a PyPI trusted publisher.
+#   2. build  — validates version/tag match, builds tiferet into
+#               a dist artifact, runs twine check.
+#   3. publish — stable or pre-release publish job, each with a dedicated
+#                GitHub environment mapped 1:1 to a PyPI trusted publisher.
 #
-# Environments (4):
+# Environments (2):
 #   pypi-tiferet             — stable releases of tiferet
 #   pypi-tiferet-pre         — pre-releases of tiferet
-#   pypi-tiferet-testing     — stable releases of tiferet-testing
-#   pypi-tiferet-testing-pre — pre-releases of tiferet-testing
 
 name: Build and Publish Tiferet
 
@@ -42,7 +40,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e . -e tiferet_testing/
+          pip install -e .
           pip install pytest pytest_env
 
       - name: Run tests
@@ -68,7 +66,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install build twine
 
-      - name: Extract versions
+      - name: Extract version
         id: get_version
         run: |
           TIFERET_VERSION=$(python -c "
@@ -76,53 +74,31 @@ jobs:
           text = pathlib.Path('tiferet/__init__.py').read_text()
           print(re.search(r\"__version__\s*=\s*['\\\"](.*?)['\\\"]\" , text).group(1))
           ")
-          TESTING_VERSION=$(python -c "
-          import re, pathlib
-          text = pathlib.Path('tiferet_testing/src/tiferet_testing/__init__.py').read_text()
-          print(re.search(r\"__version__\s*=\s*['\\\"](.*?)['\\\"]\" , text).group(1))
-          ")
           echo "tiferet_version=$TIFERET_VERSION" >> $GITHUB_OUTPUT
-          echo "testing_version=$TESTING_VERSION" >> $GITHUB_OUTPUT
 
       - name: Validate version matches release tag
         run: |
           TAG=${{ github.ref_name }}
           TIFERET_VERSION=${{ steps.get_version.outputs.tiferet_version }}
-          TESTING_VERSION=${{ steps.get_version.outputs.testing_version }}
 
           if [ "$TAG" != "v$TIFERET_VERSION" ]; then
             echo "Error: Tag ($TAG) does not match tiferet version (v$TIFERET_VERSION)"
             exit 1
           fi
 
-          if [ "$TIFERET_VERSION" != "$TESTING_VERSION" ]; then
-            echo "Error: tiferet ($TIFERET_VERSION) and tiferet-testing ($TESTING_VERSION) versions do not match"
-            exit 1
-          fi
-
       - name: Build tiferet
         run: |
-          python -m build --outdir dist-tiferet/
+          python -m build --outdir dist/
 
-      - name: Build tiferet-testing
+      - name: Check package
         run: |
-          python -m build tiferet_testing/ --outdir dist-tiferet-testing/
+          twine check dist/*
 
-      - name: Check packages
-        run: |
-          twine check dist-tiferet/* dist-tiferet-testing/*
-
-      - name: Upload tiferet artifacts
+      - name: Upload artifacts
         uses: actions/upload-artifact@v4
         with:
           name: dist-tiferet
-          path: dist-tiferet/
-
-      - name: Upload tiferet-testing artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: dist-tiferet-testing
-          path: dist-tiferet-testing/
+          path: dist/
 
   # ── tiferet stable ──────────────────────────────────────────
   publish-tiferet:
@@ -149,33 +125,5 @@ jobs:
     steps:
       - uses: actions/download-artifact@v4
         with: { name: dist-tiferet, path: dist/ }
-      - uses: pypa/gh-action-pypi-publish@release/v1
-        with: { repository-url: 'https://upload.pypi.org/legacy/' }
-
-  # ── tiferet-testing stable ──────────────────────────────────
-  publish-tiferet-testing:
-    runs-on: ubuntu-latest
-    needs: build
-    if: ${{ !contains(github.ref_name, 'a') && !contains(github.ref_name, 'b') && !contains(github.ref_name, 'rc') }}
-    environment: pypi-tiferet-testing
-    permissions:
-      id-token: write
-    steps:
-      - uses: actions/download-artifact@v4
-        with: { name: dist-tiferet-testing, path: dist/ }
-      - uses: pypa/gh-action-pypi-publish@release/v1
-        with: { repository-url: 'https://upload.pypi.org/legacy/' }
-
-  # ── tiferet-testing pre-release ─────────────────────────────
-  publish-tiferet-testing-pre:
-    runs-on: ubuntu-latest
-    needs: build
-    if: ${{ contains(github.ref_name, 'a') || contains(github.ref_name, 'b') || contains(github.ref_name, 'rc') }}
-    environment: pypi-tiferet-testing-pre
-    permissions:
-      id-token: write
-    steps:
-      - uses: actions/download-artifact@v4
-        with: { name: dist-tiferet-testing, path: dist/ }
       - uses: pypa/gh-action-pypi-publish@release/v1
         with: { repository-url: 'https://upload.pypi.org/legacy/' }
