+++ tests/events/conftest.py
@@ -3,7 +3,7 @@
 # *** imports
 
 # ** app
-from tiferet_testing.hooks import register_event_hooks
+from tiferet.testing.hooks import register_event_hooks
 
 # *** hooks
