+++ tests/mappers/test_error.py
@@ -13,7 +13,7 @@ from tiferet.mappers.error import (
     ErrorYamlObject,
     ErrorMessageYamlObject,
 )
-from tiferet_testing import AggregateTestBase, TransferObjectTestBase
+from tiferet.testing import AggregateTestBase, TransferObjectTestBase
 
 
 # *** constants
