--- tiferet_testing/pyproject.toml
@@ -1,39 +1,0 @@
-[project]
-name = "tiferet-testing"
-dynamic = ["version"]
-description = "Test harness, base classes, and helpers for building tests with the Tiferet DDD framework."
-authors = [
-    { name = "Andrew Shatz, Great Strength Systems", email = "andrew@greatstrength.me" }
-]
-license = { text = "BSD-3-Clause" }
-readme = "README.md"
-requires-python = ">=3.10"
-dependencies = [
-    "tiferet>=2.0.0b4",
-    "pytest>=8.3.3"
-]
-classifiers = [
-    "Development Status :: 4 - Beta",
-    "Framework :: Pytest",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: BSD License",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Topic :: Software Development :: Testing",
-]
-
-[project.urls]
-Homepage = "https://github.com/greatstrength/tiferet"
-Repository = "https://github.com/greatstrength/tiferet"
-
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["tiferet_testing"]
-
-[tool.setuptools.dynamic]
-version = { attr = "tiferet_testing.__version__" }
